♻️refactor/MainIcon 컴포넌트 컴포넌트화 리팩토링
